align: extract 7 inline test blocks; capture follow-up plan
Rule 8 (separate test files): the project convention is sibling
foo_tests.rs files wired in with
`#[cfg(test)] #[path = "..._tests.rs"] mod tests;`, not inline
`#[cfg(test)] mod tests { ... }` blocks. Audit found 22 files still
using the inline form. Extract the 7 smallest (each under 30 lines)
to set the pattern in this branch:

- market_data/realtime/mod.rs -> mod_tests.rs
- transport/common.rs -> common_tests.rs
- contracts/common/stream_decoders.rs -> stream_decoders_tests.rs
- display_groups/common/encoders.rs -> encoders_tests.rs
- wsh/common/stream_decoders.rs -> stream_decoders_tests.rs
- market_data/historical/common/encoders.rs -> encoders_tests.rs
- scanner/common/stream_decoders.rs -> stream_decoders_tests.rs

Each is a 1-to-1 move: the entire `use super::*;` + test fns block
is lifted into the sibling file, and the original site keeps only
the `#[cfg(test)] #[path = "..."] mod tests;` declaration. No test
behaviour changes; cargo test --features sync reports 244 passed
0 failed across the migrated suites.

The remaining 15 inline blocks (90 to 350 lines each) plus the
audit's deferred items (rule 19 too_many_arguments, rule 4 4+-arg
functions, rule 18 ~30 async pub fns missing # Examples, rule 2
5 domains still using nested {sync,async}/mod.rs layout) are
captured in plans/code-consistency-followups.md as separate
follow-up PRs. See that plan for blast radius, file lists, and
the mechanical recipe per item.
