+++ src/market_data/realtime/mod.rs
@@ -474,14 +474,5 @@ pub use sync::*;
 // Async API methods are now on Client directly via realtime/async.rs
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_what_to_show_display() {
-        assert_eq!(WhatToShow::Trades.to_string(), "TRADES");
-        assert_eq!(WhatToShow::MidPoint.to_string(), "MIDPOINT");
-        assert_eq!(WhatToShow::Bid.to_string(), "BID");
-        assert_eq!(WhatToShow::Ask.to_string(), "ASK");
-    }
-}
+#[path = "mod_tests.rs"]
+mod tests;
