+++ src/market_data/realtime/mod_tests.rs
@@ -1,0 +1,9 @@
+use super::*;
+
+#[test]
+fn test_what_to_show_display() {
+    assert_eq!(WhatToShow::Trades.to_string(), "TRADES");
+    assert_eq!(WhatToShow::MidPoint.to_string(), "MIDPOINT");
+    assert_eq!(WhatToShow::Bid.to_string(), "BID");
+    assert_eq!(WhatToShow::Ask.to_string(), "ASK");
+}
