+++ src/transport/common_tests.rs
@@ -1,0 +1,14 @@
+use super::*;
+
+#[test]
+fn test_fibonacci_backoff() {
+    let mut backoff = FibonacciBackoff::new(10);
+
+    assert_eq!(backoff.next_delay(), Duration::from_secs(1));
+    assert_eq!(backoff.next_delay(), Duration::from_secs(2));
+    assert_eq!(backoff.next_delay(), Duration::from_secs(3));
+    assert_eq!(backoff.next_delay(), Duration::from_secs(5));
+    assert_eq!(backoff.next_delay(), Duration::from_secs(8));
+    assert_eq!(backoff.next_delay(), Duration::from_secs(10)); // capped at max
+    assert_eq!(backoff.next_delay(), Duration::from_secs(10)); // stays at max
+}
